https://dev.azure.com/umbraco/D-Team%20Tracker/_workitems/edit/6587 - Moved RelationController
